move old aws-lambda example to deploy-serverless
