--- examples/aws-lambda/requirements.txt
@@ -1,1 +1,0 @@
-numpy==1.12.0
