--- examples/aws-lambda/serverless/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "bentoml-aws-lambda-example",
-  "description": "",
-  "version": "0.1.0",
-  "dependencies": {},
-  "devDependencies": {
-    "serverless-offline": "^4.9.4",
-    "serverless-python-requirements": "^4.3.0"
-  }
-}
